+++ src/libawkward/array/ListOffsetArray.cpp
@@ -355,7 +355,7 @@ namespace awkward {
       starts.offset(),
       stops.offset());
     util::handle_error(err, classname(), identities_.get());
-    std::vector<ssize_t> shape({ lenstarts });
+    std::vector<ssize_t> shape({ (ssize_t)lenstarts });
     std::vector<ssize_t> strides({ sizeof(T) });
     std::string format;
 #ifdef _MSC_VER
